rename MoviesScrollbar to Scrollbar and move to features folder
